Test commit to test and compare Sonnet based CRA with CRabbit
