+++ src/sqlfluff/core/parser/grammar/base.py
@@ -1,7 +1,6 @@
 """Base grammar, Ref, Anything and Nothing."""
 
 import copy
-from dataclasses import dataclass
 from typing import (
     TYPE_CHECKING,
     Any,
@@ -14,69 +13,21 @@ from typing import (
     Tuple,
     TypeVar,
     Union,
-    cast,
 )
 from uuid import UUID, uuid4
 
-from sqlfluff.core.errors import SQLParseError
 from sqlfluff.core.parser.context import ParseContext
 from sqlfluff.core.parser.helpers import trim_non_code_segments
-from sqlfluff.core.parser.match_logging import (
-    LateBoundJoinSegmentsCurtailed,
-    parse_match_logging,
-)
+from sqlfluff.core.parser.match_logging import parse_match_logging
 from sqlfluff.core.parser.match_result import MatchResult
 from sqlfluff.core.parser.match_wrapper import match_wrapper
 from sqlfluff.core.parser.matchable import Matchable
-from sqlfluff.core.parser.segments import BaseSegment, BracketedSegment, allow_ephemeral
+from sqlfluff.core.parser.segments import BaseSegment, allow_ephemeral
 from sqlfluff.core.parser.types import MatchableType, SimpleHintType
 from sqlfluff.core.string_helpers import curtail_string
 
 if TYPE_CHECKING:  # pragma: no cover
     from sqlfluff.core.dialects.base import Dialect
-
-
-def first_trimmed_raw(seg: BaseSegment) -> str:
-    """Trim whitespace off a whole element raw.
-
-    Used as a helper function in BaseGrammar._look_ahead_match.
-
-    For existing compound segments, we should assume that within
-    that segment, things are internally consistent, that means
-    rather than enumerating all the individual segments of a longer
-    one we just dump out the whole segment, but splitting off the
-    first element separated by whitespace. This is a) faster and
-    also b) prevents some really horrible bugs with bracket matching.
-    See https://github.com/sqlfluff/sqlfluff/issues/433
-
-    This fetches the _whole_ raw of a potentially compound segment
-    to match against, trimming off any whitespace. This is the
-    most efficient way to get at the first element of a potentially
-    longer segment.
-    """
-    s = seg.raw_upper.split(maxsplit=1)
-    return s[0] if s else ""
-
-
-@dataclass
-class BracketInfo:
-    """BracketInfo tuple for keeping track of brackets during matching.
-
-    This is used in BaseGrammar._bracket_sensitive_look_ahead_match but
-    defined here for type checking.
-    """
-
-    bracket: BaseSegment
-    segments: Tuple[BaseSegment, ...]
-
-    def to_segment(self, end_bracket: Tuple[BaseSegment, ...]) -> BracketedSegment:
-        """Turn the contained segments into a bracketed segment."""
-        assert len(end_bracket) == 1
-        return BracketedSegment(
-            segments=self.segments,
-            start_bracket=(self.bracket,),
-            end_bracket=cast(Tuple[BaseSegment], end_bracket),
-        )
 
 
 def cached_method_for_parse_context(
@@ -520,355 +471,6 @@ class BaseGrammar(Matchable):
         # If no match at all, return nothing
         return MatchResult.from_unmatched(segments), None
 
-    @classmethod
-    def _look_ahead_match(
-        cls,
-        segments: Tuple[BaseSegment, ...],
-        matchers: List[MatchableType],
-        parse_context: ParseContext,
-    ) -> Tuple[Tuple[BaseSegment, ...], MatchResult, Optional[MatchableType]]:
-        """Look ahead for matches beyond the first element of the segments list.
-
-        This function also contains the performance improved hash-matching approach to
-        searching for matches, which should significantly improve performance.
-
-        Prioritise the first match, and if multiple match at the same point the longest.
-        If two matches of the same length match at the same time, then it's the first in
-        the iterable of matchers.
-
-        Returns:
-            `tuple` of (unmatched_segments, match_object, matcher).
-
-        """
-        parse_match_logging(
-            cls.__name__,
-            "_look_ahead_match",
-            "IN",
-            parse_context=parse_context,
-            v_level=4,
-            ls=len(segments),
-            seg=LateBoundJoinSegmentsCurtailed(segments),
-        )
-
-        # Have we been passed an empty tuple?
-        if not segments:  # pragma: no cover TODO?
-            return ((), MatchResult.from_empty(), None)
-
-        # Here we enable a performance optimisation. Most of the time in this cycle
-        # happens in loops looking for simple matchers which we should
-        # be able to find a shortcut for.
-
-        parse_match_logging(
-            cls.__name__,
-            "_look_ahead_match",
-            "SI",
-            parse_context=parse_context,
-            v_level=4,
-        )
-
-        best_simple_match = None
-        simple_match = None
-        for idx, seg in enumerate(segments):
-            trimmed_seg = first_trimmed_raw(seg)
-            for matcher in matchers:
-                simple_match = None
-                simple = matcher.simple(parse_context=parse_context)
-                if not simple:  # pragma: no cover
-                    # NOTE: For all bundled dialects, this clause is true, but until
-                    # the RegexMatcher is completely deprecated (and therefore that
-                    # `.simple()` must provide a result), it is still _possible_
-                    # to end up here.
-                    raise NotImplementedError(
-                        "All matchers passed to `._look_ahead_match()` are "
-                        "assumed to have a functioning `.simple()` option. "
-                        "In a future release it will be compulsory for _all_ "
-                        "matchables to implement `.simple()`. Please report "
-                        "this as a bug on GitHub along with your current query "
-                        f"and dialect.\nProblematic matcher: {matcher}"
-                    )
-                simple_raws, simple_types = simple
-
-                assert simple_raws or simple_types
-                if simple_raws:
-                    if trimmed_seg in simple_raws:
-                        simple_match = matcher
-
-                if simple_types and not simple_match:
-                    intersection = simple_types.intersection(seg.class_types)
-                    if intersection:
-                        simple_match = matcher
-
-                # If we couldn't achieve a simple match, move on to the next option.
-                if not simple_match:
-                    continue
-
-                # If there is, check the full version matches. If it doesn't
-                # then discount it and move on.
-                match = simple_match.match(segments[idx:], parse_context)
-                if not match:
-                    continue
-
-                best_simple_match = (
-                    segments[:idx],
-                    match,
-                    simple_match,
-                )
-                # Stop looking through matchers
-                break
-
-            # If we have a valid match, stop looking through segments
-            if best_simple_match:
-                break
-
-        # There are no other matchers, we can just shortcut now. Either with
-        # no match, or the best one we found (if we found one).
-        parse_match_logging(
-            cls.__name__,
-            "_look_ahead_match",
-            "SC",
-            parse_context=parse_context,
-            v_level=4,
-            bsm=None
-            if not best_simple_match
-            else (
-                len(best_simple_match[0]),
-                len(best_simple_match[1]),
-                best_simple_match[2],
-            ),
-        )
-
-        if best_simple_match:
-            return best_simple_match
-        else:
-            return ((), MatchResult.from_unmatched(segments), None)
-
-    @classmethod
-    def _bracket_sensitive_look_ahead_match(
-        cls,
-        segments: Tuple[BaseSegment, ...],
-        matchers: List[MatchableType],
-        parse_context: ParseContext,
-        start_bracket: Optional[MatchableType] = None,
-        end_bracket: Optional[MatchableType] = None,
-        bracket_pairs_set: str = "bracket_pairs",
-    ) -> Tuple[Tuple[BaseSegment, ...], MatchResult, Optional[MatchableType]]:
-        """Same as `_look_ahead_match` but with bracket counting.
-
-        NB: Given we depend on `_look_ahead_match` we can also utilise
-        the same performance optimisations which are implemented there.
-
-        bracket_pairs_set: Allows specific segments to override the available
-            bracket pairs. See the definition of "angle_bracket_pairs" in the
-            BigQuery dialect for additional context on why this exists.
-
-        Returns:
-            `tuple` of (unmatched_segments, match_object, matcher).
-
-        """
-        # Have we been passed an empty tuple?
-        if not segments:
-            return ((), MatchResult.from_unmatched(segments), None)
-
-        # Get hold of the bracket matchers from the dialect, and append them
-        # to the list of matchers. We get them from the relevant set on the
-        # dialect. We use zip twice to "unzip" them. We ignore the first
-        # argument because that's just the name.
-        _, start_bracket_refs, end_bracket_refs, persists = zip(
-            *parse_context.dialect.bracket_sets(bracket_pairs_set)
-        )
-        # These are matchables, probably StringParsers.
-        start_brackets = [
-            parse_context.dialect.ref(seg_ref) for seg_ref in start_bracket_refs
-        ]
-        end_brackets = [
-            parse_context.dialect.ref(seg_ref) for seg_ref in end_bracket_refs
-        ]
-        # Add any bracket-like things passed as arguments
-        if start_bracket:
-            start_brackets += [start_bracket]
-        if end_bracket:
-            end_brackets += [end_bracket]
-        bracket_matchers = start_brackets + end_brackets
-
-        # Make some buffers
-        seg_buff: Tuple[BaseSegment, ...] = segments
-        pre_seg_buff: Tuple[BaseSegment, ...] = ()
-        bracket_stack: List[BracketInfo] = []
-
-        # Iterate
-        while True:
-            # Do we have anything left to match on?
-            if seg_buff:
-                # Yes we have buffer left to work with.
-                # Are we already in a bracket stack?
-                if bracket_stack:
-                    # Yes, we're just looking for the closing bracket, or
-                    # another opening bracket.
-                    pre, match, matcher = cls._look_ahead_match(
-                        seg_buff,
-                        bracket_matchers,
-                        parse_context=parse_context,
-                    )
-
-                    if match:
-                        # NB: We can only consider this as a nested bracket if the start
-                        # and end tokens are not the same. If a matcher is both a start
-                        # and end token we cannot deepen the bracket stack. In general,
-                        # quoted strings are a typical example where the start and end
-                        # tokens are the same. Currently, though, quoted strings are
-                        # handled elsewhere in the parser, and there are no cases where
-                        # *this* code has to handle identical start and end brackets.
-                        # For now, consider this a small, speculative investment in a
-                        # possible future requirement.
-                        if matcher in start_brackets and matcher not in end_brackets:
-                            # Add any segments leading up to this to the previous
-                            # bracket.
-                            bracket_stack[-1].segments += pre
-                            # Add a bracket to the stack and add the matches from the
-                            # segment.
-                            bracket_stack.append(
-                                BracketInfo(
-                                    bracket=match.matched_segments[0],
-                                    segments=match.matched_segments,
-                                )
-                            )
-                            seg_buff = match.unmatched_segments
-                            continue
-                        elif matcher in end_brackets:
-                            # Found an end bracket. Does its type match that of
-                            # the innermost start bracket? E.g. ")" matches "(",
-                            # "]" matches "[".
-                            # For the start bracket we don't have the matcher
-                            # but we can work out the type, so we use that for
-                            # the lookup.
-                            start_index = [
-                                bracket.type for bracket in start_brackets
-                            ].index(bracket_stack[-1].bracket.get_type())
-                            # For the end index, we can just look for the matcher
-                            end_index = end_brackets.index(matcher)
-                            bracket_types_match = start_index == end_index
-                            if bracket_types_match:
-                                # Yes, the types match. So we've found a
-                                # matching end bracket. Pop the stack, construct
-                                # a bracketed segment and carry
-                                # on.
-
-                                # Complete the bracketed info
-                                bracket_stack[-1].segments += (
-                                    pre + match.matched_segments
-                                )
-                                # Construct a bracketed segment (as a tuple) if allowed.
-                                persist_bracket = persists[end_brackets.index(matcher)]
-                                if persist_bracket:
-                                    new_segments: Tuple[BaseSegment, ...] = (
-                                        bracket_stack[-1].to_segment(
-                                            end_bracket=match.matched_segments
-                                        ),
-                                    )
-                                else:
-                                    new_segments = bracket_stack[-1].segments
-                                # Remove the bracket set from the stack
-                                bracket_stack.pop()
-                                # If we're still in a bracket, add the new segments to
-                                # that bracket, otherwise add them to the buffer
-                                if bracket_stack:
-                                    bracket_stack[-1].segments += new_segments
-                                else:
-                                    pre_seg_buff += new_segments
-                                seg_buff = match.unmatched_segments
-                                continue
-                            else:
-                                # The types don't match. Error.
-                                raise SQLParseError(
-                                    f"Found unexpected end bracket!, "
-                                    f"was expecting "
-                                    f"{end_brackets[start_index]}, "
-                                    f"but got {matcher}",
-                                    segment=match.matched_segments[0],
-                                )
-
-                        else:  # pragma: no cover
-                            raise RuntimeError("I don't know how we get here?!")
-                    else:  # pragma: no cover
-                        # No match, we're in a bracket stack. Error.
-                        raise SQLParseError(
-                            "Couldn't find closing bracket for opening bracket.",
-                            segment=bracket_stack[-1].bracket,
-                        )
-                else:
-                    # No, we're open to more opening brackets or the thing(s)
-                    # that we're otherwise looking for.
-                    pre, match, matcher = cls._look_ahead_match(
-                        seg_buff,
-                        matchers + bracket_matchers,
-                        parse_context=parse_context,
-                    )
-
-                    if match:
-                        if matcher in matchers:
-                            # It's one of the things we were looking for!
-                            # Return.
-                            return (pre_seg_buff + pre, match, matcher)
-                        elif matcher in start_brackets:
-                            # We've found the start of a bracket segment.
-                            # NB: It might not *actually* be the bracket itself,
-                            # but could be some non-code element preceding it.
-                            # That's actually ok.
-
-                            # Add the bracket to the stack.
-                            bracket_stack.append(
-                                BracketInfo(
-                                    bracket=match.matched_segments[0],
-                                    segments=match.matched_segments,
-                                )
-                            )
-                            # The matched element has already been added to the bracket.
-                            # Add anything before it to the pre segment buffer.
-                            # Reset the working buffer.
-                            pre_seg_buff += pre
-                            seg_buff = match.unmatched_segments
-                            continue
-                        elif matcher in end_brackets:
-                            # We've found an unexpected end bracket! This is likely
-                            # because we're matching a section which should have ended.
-                            # If we had a match, it would have matched by now, so this
-                            # means no match.
-                            parse_match_logging(
-                                cls.__name__,
-                                "_bracket_sensitive_look_ahead_match",
-                                "UEXB",
-                                parse_context=parse_context,
-                                v_level=3,
-                                got=matcher,
-                            )
-                            # From here we'll drop out to the happy unmatched exit.
-                        else:  # pragma: no cover
-                            # This shouldn't happen!?
-                            raise NotImplementedError(
-                                "This shouldn't happen. Panic in "
-                                "_bracket_sensitive_look_ahead_match."
-                            )
-                    # Not in a bracket stack, but no match.
-                    # From here we'll drop out to the happy unmatched exit.
-            else:
-                # No we're at the end:
-                # Now check have we closed all our brackets?
-                if bracket_stack:  # pragma: no cover
-                    # No we haven't.
-                    raise SQLParseError(
-                        "Couldn't find closing bracket for opened brackets: "
-                        f"`{bracket_stack}`.",
-                        segment=bracket_stack[-1].bracket,
-                    )
-
-            # This is the happy unmatched path. This occurs when:
-            # - We reached the end with no open brackets.
-            # - No match while outside a bracket stack.
-            # - We found an unexpected end bracket before matching something
-            # interesting. We return with the mutated segments so we can reuse any
-            # bracket matching.
-            return ((), MatchResult.from_unmatched(pre_seg_buff + seg_buff), None)
-
     def __str__(self) -> str:  # pragma: no cover TODO?
         return repr(self)
 
@@ -1052,7 +654,7 @@ class Ref(BaseGrammar):
 
     def __repr__(self) -> str:
         return "<Ref: {}{}>".format(
-            str(self._ref), " [opt]" if self.is_optional() else ""
+            repr(self._ref), " [opt]" if self.is_optional() else ""
         )
 
     @match_wrapper(v_level=4)  # Log less for Ref
