+++ src/sqlfluff/core/parser/match_logging.py
@@ -1,12 +1,9 @@
 """Classes to help with match logging."""
 
 import logging
-from typing import TYPE_CHECKING, Any, Tuple
-
-from sqlfluff.core.parser.helpers import join_segments_raw_curtailed
+from typing import TYPE_CHECKING, Any
 
 if TYPE_CHECKING:  # pragma: no cover
-    from sqlfluff.core.parser import BaseSegment
     from sqlfluff.core.parser.context import ParseContext
 
 
@@ -100,17 +97,3 @@ def parse_match_logging(
     ParseMatchLogObject(
         parse_context, grammar, func, msg, v_level=v_level, **kwargs
     ).log()
-
-
-class LateBoundJoinSegmentsCurtailed:
-    """Object to delay `join_segments_raw_curtailed` until later.
-
-    This allows us to defer the string manipulation involved
-    until actually required by the logger.
-    """
-
-    def __init__(self, segments: Tuple["BaseSegment", ...]) -> None:
-        self.segments = segments
-
-    def __str__(self) -> str:
-        return repr(join_segments_raw_curtailed(self.segments))
